refactor(prompt-compliance): move models
