--- src/a2a_t/server/prompt_compliance/result.py
@@ -1,11 +1,0 @@
-from __future__ import annotations
-
-from dataclasses import dataclass
-
-
-@dataclass
-class PromptComplianceResult:
-    """Unified compliance execution result."""
-
-    success: bool
-    failure: dict[str, str] | None = None
