+++ tests/prompt_compliance/test_a2at_server.py
@@ -15,7 +15,7 @@ if str(SRC_ROOT) not in sys.path:
 
 from a2a_t.negotiation.common.enums import NegotiationStatus, NegotiationType
 from a2a_t.negotiation.common.models import ContinueNegotiationInput, NegotiationContext, StartNegotiationInput
-from a2a_t.server.prompt_compliance.result import PromptComplianceResult
+from a2a_t.server.prompt_compliance.models import PromptComplianceResult
 from tests.test_support import ManagedTempDirTestCase, TEST_ENV_PATH
